UnitObject and CityObject subClasses - ALPHA 1.1.5 Minor update to refactor RUnit and City to ensure that view controls are in a subclass called RUnitObject and CityObject
Requirement:
There's been a lot of problems brought about by the fact that RUnit has view controls buut is also abstract (i.e. you can't implement view changes in the RUnit Dictionary in Game, as it is not an instance in the scene.

If we therefore have a sub-class that handles the movement etc things would be better. We could then undertake a massive overhaul and make the whole more thing object oriented, with far less of the actual actions taking place in gameController, and it reverting back to a controller rather than a place where actions are defined.

**FEATURE IMPLEMENTED - ALPHA 1.1.5**
Relatively simple to implement, just created subclass called UnitObject which inherits from RUnit and changed relevant scripts in RGameController.

Primary Scripts affected:
- RGameController - SyncSceneUnitwithDicitonaryUnit
- RGameController - OnBoardClicked - (switched out GetComponent<RUnit> with GetComponent<UnitObject> where approrpriate
- UIController - Split - as above, switched out GetComponent<RUnit> with <UnitObject>